Move schedule_calcs.hpp functions into multiple ofther scheduler files.
