+++ speeduino/scheduler.cpp
@@ -31,8 +31,6 @@ A full copy of the license may be found in the projects root directory
 #include "units.h"
 #include "schedule_state_machine.h"
 #include "unit_testing.h"
-#include "decoders.h"
-#include "scheduledIO_inj.h"
 
 void nullCallback(void) { return; }
 
@@ -182,3 +180,29 @@ void moveToNextState(IgnitionSchedule &schedule)  noexcept
   movetoNextState(schedule, ignitionPendingToRunning, ignitionRunningToOff, ignitionRunningToPending);
 }
 
+void adjustCrankAngle(const statuses &current, IgnitionSchedule &schedule, int16_t crankAngle) {
+  constexpr uint8_t MIN_CYCLES_FOR_CORRECTION = 6U;
+
+  crankAngle = ignitionLimits(crankAngle);
+  ATOMIC() { // Prevent race conditions with the timer interrupt.
+    // We only want to adjust the crank angle if we are running and the coil is charging or we are waiting for the timer to fire.
+    if( isRunning(schedule) ) {
+      if  (schedule.dischargeAngle>crankAngle) { 
+        // Coil is charging so change the charge time so the spark fires at
+        // the requested crank angle (this could reduce dwell time & potentially
+        // result in a weaker spark).
+        SET_COMPARE(schedule._compare, schedule._counter + angleToTimerTicks( schedule.dischargeAngle-crankAngle )); 
+      } 
+    }
+    else if( (schedule._status==PENDING) ) {
+      if ((current.startRevolutions > MIN_CYCLES_FOR_CORRECTION) && (schedule.chargeAngle>crankAngle)) {
+        // We are waiting for the timer to fire & start charging the coil.
+        // Keep dwell (I.e. duration) constant (for better spark) - instead adjust the waiting period so 
+        // the spark fires at the requested crank angle.
+        SET_COMPARE(schedule._compare, schedule._counter + angleToTimerTicks( schedule.chargeAngle-crankAngle )); 
+      }
+    } else {
+      // Unknown state, so no adjustment possible
+    }
+  }
+}
