+++ test/test_decoders/NGC/test_ngc.cpp
@@ -6,6 +6,7 @@
 #include "scheduler_ignition_controller.h"
 
 extern uint16_t ignitionEndTeeth[IGN_CHANNELS];
+extern void calculateIgnitionAngles(IgnitionSchedule &schedule, uint16_t dwellAngle, int8_t advance);
 
 void test_ngc_newIgn_12_trig0_1()
 {
