+++ test/test_schedule_calcs/test_adjust_crank_angle.cpp
@@ -1,9 +1,8 @@
-#include <Arduino.h>
 #include <unity.h>
 #include "scheduler.h"
 #include "../test_utils.h"
 #include "src/stdlib/type_traits.h"
-
+#include "statuses.h"
 
 using raw_counter_t = type_traits::remove_reference<IgnitionSchedule::counter_t>::type;
 using raw_compare_t = type_traits::remove_reference<IgnitionSchedule::compare_t>::type;
